refactor: rename auth/components to auth/pages
